feat(tokens): move chakra formartting function to utils file/folder, add comments documenting chakra format, add comments documenting ts literal types in generated/ts/index.ts, add README, restructure to use src folder, update imports after restrucrturing files
