+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-monorepo",
+  "name": "design-system-monorepo",
   "private": true,
   "workspaces": [
     "packages/*",
