+++ packages/bleh-ui/src/theme/tokens/line-heights.ts
@@ -1,4 +1,4 @@
 import { defineTokens } from "@chakra-ui/react";
-import designTokens from "../../../../tokens/generated/chakra/theme-tokens";
+import designTokens from "../../../../tokens/src/generated/chakra/theme-tokens";
 
 export const lineHeights = defineTokens.lineHeights(designTokens.lineHeight);
