+++ packages/bleh-ui/src/theme/tokens/radii.ts
@@ -1,4 +1,4 @@
 import { defineTokens } from "@chakra-ui/react";
-import designTokens from "../../../../tokens/generated/chakra/theme-tokens";
+import designTokens from "../../../../tokens/src/generated/chakra/theme-tokens";
 
 export const radii = defineTokens.radii(designTokens.borderRadius);
