+++ packages/tokens/package.json
@@ -3,13 +3,14 @@
   "version": "0.0.1",
   "type": "module",
   "scripts": {
-    "build": "style-dictionary build --config ./style-dictionary.config.js"
+    "build": "style-dictionary build --config ./src/style-dictionary.config.js"
   },
   "packageManager": "pnpm@9.12.3+sha512.cce0f9de9c5a7c95bef944169cc5dfe8741abfb145078c0d508b868056848a87c81e626246cb60967cbd7fd29a6c062ef73ff840d96b3c86c40ac92cf4a813ee",
   "devDependencies": {
     "@tokens-studio/sd-transforms": "^1.2.9",
     "is-plain-obj": "^4.1.0",
     "prettier": "3.4.2",
-    "style-dictionary": "^4.3.0"
+    "style-dictionary": "^4.3.0",
+    "ts-node": "^10.9.2"
   }
 }
